requirements.txt
python-dotenv
requests
pandas
matplotlib
seaborn
plotly
fastapi
uvicorn
duckdb
anthropic
